+++ src/term.c
@@ -3205,7 +3205,7 @@ emoji_show(int x, int y, struct emoji e, int elen, cattr eattr, ushort lattr)
 #ifdef debug_win_text_invocation
 
 void
-_win_text(int line, int tx, int ty, wchar *text, int len, cattr attr, cattr *textattr, ushort lattr, char has_rtl, bool clearpad, uchar phase)
+_win_text(int line, int tx, int ty, wchar *text, int len, cattr attr, cattr *textattr, ushort lattr, char has_rtl, char has_sea, bool clearpad, uchar phase)
 {
   if (*text != ' ') {
     printf("[%d] %d:%d(len %d) attr %08llX", line, ty, tx, len, attr.attr);
@@ -3213,10 +3213,10 @@ _win_text(int line, int tx, int ty, wchar *text, int len, cattr attr, cattr *tex
       printf(" %04X", text[i]);
     printf("\n");
   }
-  win_text(tx, ty, text, len, attr, textattr, lattr, has_rtl, clearpad, phase);
+  win_text(tx, ty, text, len, attr, textattr, lattr, has_rtl, has_sea, clearpad, phase);
 }
 
-#define win_text(tx, ty, text, len, attr, textattr, lattr, has_rtl, clearpad, phase) _win_text(__LINE__, tx, ty, text, len, attr, textattr, lattr, has_rtl, clearpad, phase)
+#define win_text(tx, ty, text, len, attr, textattr, lattr, has_rtl, has_sea, clearpad, phase) _win_text(__LINE__, tx, ty, text, len, attr, textattr, lattr, has_rtl, has_sea, clearpad, phase)
 
 #endif
 
@@ -3248,6 +3248,23 @@ void trace_line(char * tag, termchar * chars)
 
 #define IGNWIDTH TATTR_EXPAND | TATTR_NARROW | TATTR_SINGLE | TATTR_CLEAR
 #define IGNEMOJATTR (TATTR_WIDE | ATTR_FGMASK | TATTR_COMBINING | IGNWIDTH)
+
+static bool
+is_comcom(wchar ch)
+{
+  static wchar comcom[] = {
+    0x0E33,  // THAI CHARACTER SARA AM
+    0x0EB3,  // LAO VOWEL SIGN AM
+    0x0F77,  // TIBETAN VOWEL SIGN VOCALIC RR
+    0x0F79,  // TIBETAN VOWEL SIGN VOCALIC LL
+  };
+  for (uint i = 0; i < lengthof(comcom); i++)
+    if (ch < comcom[i])
+      return false;
+    else if (ch == comcom[i])
+      return true;
+  return false;
+}
 
 void
 (term_paint)(struct term* term_p)
@@ -4123,11 +4140,11 @@ void
     }
     if (prevdirtyitalic) {
       // clear overhang into right padding border
-      win_text(term.cols, i, const_cast<wchar *>(W(" ")), 1, CATTR_DEFAULT, (cattr*)&CATTR_DEFAULT, line->lattr, false, true, 0);
+      win_text(term.cols, i, const_cast<wchar *>(W(" ")), 1, CATTR_DEFAULT, (cattr*)&CATTR_DEFAULT, line->lattr, false, false, true, 0);
     }
     if (firstdirtyitalic) {
       // clear overhang into left padding border
-      win_text(-1, i, const_cast<wchar *>(W(" ")), 1, CATTR_DEFAULT, (cattr*)&CATTR_DEFAULT, line->lattr, false, true, 0);
+      win_text(-1, i, const_cast<wchar *>(W(" ")), 1, CATTR_DEFAULT, (cattr*)&CATTR_DEFAULT, line->lattr, false, false, true, 0);
     }
 
 #define dont_debug_bidi_paragraphs
@@ -4172,6 +4189,7 @@ void
     int textlen = 0;
 
     char has_rtl = 0;
+    char has_sea = 0;  // South East Asian script
     uchar bc = 0;
     bool dirty_run = (line->lattr != displine->lattr);
     bool dirty_line = dirty_run;
@@ -4194,11 +4212,12 @@ void
     cattr ovl_attr;
     ushort ovl_lattr;
     char ovl_has_rtl;
+    char ovl_has_sea;
 
     auto flush_text = [&]()
     {
       if (ovl_len) {
-        win_text(ovl_x, ovl_y, ovl_text, ovl_len, ovl_attr, ovl_textattr, ovl_lattr, ovl_has_rtl, false, 2);
+        win_text(ovl_x, ovl_y, ovl_text, ovl_len, ovl_attr, ovl_textattr, ovl_lattr, ovl_has_rtl, ovl_has_sea, false, 2);
         ovl_len = 0;
       }
     };
@@ -4211,7 +4230,7 @@ void
 # define debug_out_text
 #endif
 
-    auto out_text = [&](int x, int y, wchar *text, int len, cattr attr, cattr *textattr, ushort lattr, char has_rtl)
+    auto out_text = [&](int x, int y, wchar *text, int len, cattr attr, cattr *textattr, ushort lattr, char has_rtl, char has_sea)
     {
 #ifdef debug_out_text
       wchar t[len + 1]; wcsncpy(t, text, len); t[len] = 0;
@@ -4257,7 +4276,7 @@ void
             if (elen == 1 && attr.attr & TATTR_OVERHANG)
               elen = 2;
             // fill emoji background
-            win_text(x, y, esp, elen, eattr, textattr, lattr, has_rtl, false, 1);
+            win_text(x, y, esp, elen, eattr, textattr, lattr, has_rtl, has_sea, false, 1);
             flush_text();
           }
 #if defined(debug_emojis) && debug_emojis > 4
@@ -4265,7 +4284,7 @@ void
           eattr.attr &= ~(ATTR_BGMASK | ATTR_FGMASK);
           eattr.attr |= 6 << ATTR_BGSHIFT | 4;
           esp[0] = '0' + elen;
-          win_text(x, y, esp, elen, eattr, textattr, lattr, has_rtl, false, 2);
+          win_text(x, y, esp, elen, eattr, textattr, lattr, has_rtl, has_sea, false, 2);
 #endif
           if (cfg.emoji_placement == EMPL_FULL && !overlaying)
             do_overlay = true;  // display in overlaying loop
@@ -4280,7 +4299,7 @@ void
           eattr.attr &= ~(ATTR_BGMASK | ATTR_FGMASK);
           eattr.attr |= 4 << ATTR_BGSHIFT | 6;
           esp[0] = '0';
-          win_text(x, y, esp, 1, eattr, textattr, lattr, has_rtl, false, 2);
+          win_text(x, y, esp, 1, eattr, textattr, lattr, has_rtl, has_sea, false, 2);
         }
 #endif
       }
@@ -4306,7 +4325,7 @@ void
            * combining doubles
            * cursor position, to support underlay cursor painting
          */
-        win_text(x, y, text, len, attr, textattr, lattr, has_rtl, false, 1);
+        win_text(x, y, text, len, attr, textattr, lattr, has_rtl, has_sea, false, 1);
         flush_text();
         ovl_x = x;
         ovl_y = y;
@@ -4316,9 +4335,10 @@ void
         memcpy(ovl_textattr, textattr, len * sizeof(cattr));
         ovl_lattr = lattr;
         ovl_has_rtl = has_rtl;
+        ovl_has_sea = has_sea;
       }
       else {
-        win_text(x, y, text, len, attr, textattr, lattr, has_rtl, false, 0);
+        win_text(x, y, text, len, attr, textattr, lattr, has_rtl, has_sea, false, 0);
         flush_text();
       }
     };
@@ -4356,7 +4376,7 @@ void
         dirty_line = true;
 
 #ifdef debug_run
-#define trace_run(tag)	({/*if (tchar & 0xFF00)*/ if (tchar != ' ') printf("break (%s) (%04X)%04X\n", tag, j > 0 ? newchars[j - 1].chr : 0, tchar);})
+#define trace_run(tag)	({if (tchar > '~') printf("break (%s) @%d %04X|%04X\n", tag, j, j > 0 ? newchars[j - 1].chr : 0, tchar);})
 
 #else
 #define trace_run(tag)	(void)0
@@ -4427,6 +4447,22 @@ void
           trace_run("len"), break_run = true;
       }
 
+     /*
+      * Output South East Asian script combining characters in one chunk.
+      */
+      if (tchar >= 0x0900 && tchar <= 0x109F) {
+        // South Asian (U+0900-U+0DFF)
+        // South East Asian (U+0E00-U+109F)
+        has_sea |= 1;
+      }
+      if (break_run && is_comcom(tchar)) {
+        // Output composed combining characters with previous in one chunk;
+        // also extend rendering box.
+        has_sea |= 2;
+        trace_run("no-break-run");
+        break_run = false;
+      }
+
       uchar tbc = bidi_class(xtchar);
 
       if (textlen && tbc != bc) {
@@ -4446,10 +4482,11 @@ void
      /* Flush previous output chunk on break_run */
       if (break_run || cfg.bloom) {
         if ((dirty_run && textlen) || overlaying)
-          out_text(start, i, text, textlen, attr, textattr, line->lattr, has_rtl);
+          out_text(start, i, text, textlen, attr, textattr, line->lattr, has_rtl, has_sea);
         start = j;
         textlen = 0;
         has_rtl = 0;
+        has_sea = 0;
         attr = tattr;
         dirty_run = dirty_line;
 #if defined(debug_dirty) && debug_dirty > 1
@@ -4592,7 +4629,7 @@ void
       }
     }
     if (dirty_run && textlen)
-      out_text(start, i, text, textlen, attr, textattr, line->lattr, has_rtl);
+      out_text(start, i, text, textlen, attr, textattr, line->lattr, has_rtl, has_sea);
     if (!overlaying)
       flush_text();